Mostly bug fixes
The JOIN command must response with first the JOIN reply (user joining the
channel, sent to everyone including the sender) and followed by the name list.
The name list's format is:

    :{server_host} 353 {user_nick} = {channel} :{user1} {user2} {user3} ...

The `=` character denotes it is a public channel.

This commit also bug fixes several leftover nasty bugs (including a `__getattr__`
infinite recursion).

Co-authored-by: Astremy <astremy.com@gmail.com>