ref: aioircd v2
The server have been re-implemented from ground 0, here are the few main
changes:

* Use Trio over Asyncio
* Split the single aioircd.py module-file into a full-fledged aioircd/
  module-dir with one file per abstract class.
* Reimplement the IRC protocol using https://modern.ircdocs.horse
  instead of the deprecated RFC 282[1-4].
* Configuration moved from the cli to the environ vars.

Why Trio ?
==========

The former asyncio implementation suffered from the limitations of the
api. In many places we had to be clever in order to keep track of every
created task and await them all upon "main task" termination to re-link
the task's callstack in the current callstack and correctly gather the
exceptions. Basically we re-implemented parts of the Trio API to cope
with Asyncio's limitations. Thanks to its "nursery" mechanism, the Trio
API offers out-of-the-box support to the approach we implemented in the
previous version.

Co-authored-by: Antoine <antoine@enimaloc.fr>