ref: new test framework
Each key=value pair from the configuration is no longer stored in the
global scope of the `aioircd.config` module. The pairs are instead
stored in a singleton object `config` saved in the global scope. With
the previous way of storing the key=value pairs it was possible to
from-import a specific pair making it pretty hard to mock in unittests.
Using this `config` object namespace it is now possible to easily mock
the configuration is the tests.

The test framework have been enriched with many handy methods. First and
foremost we now ship an `AsyncTestCase` that takes care of creating a
`unittest`-compatible methods out of an `async def test_` methods. This
allow to easily add tests where all the trio boilerplates are setup.

The other important additions are the many IRC-related methods that have
been added on both the `FakeUser` class and on a new `TestIRC` class
case. It is now much easier to connect and register users, send/receive
messages as those users, wait for the state of a user to change
server-side.