fix: ignore missing token in PONG
Some nasty IRC client softwares send `b"PONG\r\n"` without token as
reply to a b"PING 1234\r\n". They don't properly implement the
specification the right way but we want to be compatible.

Because we actually don't verify that the token present in the original
PING command matches with the token present in the PONG reply, we
decided to soften the requirement to also accept PONG reply with no
token.