misc: CI/CD via github action
We forgot to update the README with the v2. We took the opportunity to
update the CLI to include --version and --help.

The **Lint and Test** `testing.yml` workflow runs each time a PR is
open and each time commits are push to a PR. This workflow lints the
code to find errors such as NameError. This workflow also runs the
unittests. In case any of the two fail, the PR is automatically blocked.

The **Package and Publish** `packaging.yml` workflow runs only when
commits are pushed to the master branch (e.g. when a PR is merged). This
workflow opens a new PR that change the version and update the
changelog, this PR is automatically updated until it is merge. Once
merged, new artifacts are created and uploaded to pypi.