fix: concurrent writes
It was possible to run into a race condition when multiple messages
would be sent *at the same time* to the same user. Trio doesn't
implement any kind of queue itself (which is lame, IMHO) so we
introduced a lock.

Quick note, locks in a single-thread asynchronous environment are
both VERY cheap and quite safe.

Closes: #17